Add Link unit tests
Additionally, fix a warning, which was issued from Styled-Components, when a component was styled dynamically from within another component. The solution I've found was to use the `as` property to set the rendered DOM element, whilst keeping the styling of a styled component that was already defined before, outside the function. Not sure, if this is the "best" way to do it, but it seems to work for the given scenarios.